requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
streamlit>=1.36.0
httpx>=0.27.0
openai>=1.35.0
pydantic>=2.7.0
python-dotenv>=1.0.1
